added functions to set register bits

+++ src/bits/comparator_mode.rs
@@ -5,8 +5,8 @@ pub enum ComparatorMode {
     CompModeWindow = 0b1,
 }
 impl ComparatorMode {
-    pub const fn get_value(self) -> u8 {
-        self as u8
+    pub const fn get_value(self) -> u16 {
+        self as u16
     }
 }
 impl Default for ComparatorMode {
