+++ src/bits/comparator_polarity.rs
@@ -6,8 +6,8 @@ pub enum ComparatorPolarity {
 }
 
 impl ComparatorPolarity {
-    pub const fn get_value(self) -> u8 {
-        self as u8
+    pub const fn get_value(self) -> u16 {
+        self as u16
     }
 }
