+++ src/bits/gain_amp.rs
@@ -10,8 +10,8 @@ pub enum GainAmp {
 }
 
 impl GainAmp {
-    pub const fn get_value(self) -> u8 {
-        self as u8
+    pub const fn get_value(self) -> u16 {
+        self as u16
     }
 }
